Move `SyncSet` to its own module
